*render* new: - Windows! Display device now only works as a relay to handle windows, but windows themselves are instantiateable. - Every piece of code using DisplayDevice::GetDisplayMode has been changed to fetch either the current window, or window 0 (which is the window owning the GL context).

+++ code/foundation/io/binaryreader.h
@@ -60,9 +60,9 @@ public:
     /// read a 32-bit unsigned int from the stream
     unsigned int ReadUInt();
     /// read a 64-bit int from the stream
-    int64_t ReadInt64();
+	long long ReadInt64();
     /// read a 64-bit unsigned int from the stream
-    uint64_t ReadUInt64();
+	unsigned long long ReadUInt64();
     /// read a float value from the stream
     float ReadFloat();
     /// read a compressed float value from the stream, lossy and needed to be in the range of -1.0 and +1.0
